refactor(sdk): inline cache api at crate root
